requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
jinja2==3.1.6
python-multipart==0.0.20
python-docx==1.2.0
openpyxl==3.1.5
PyMuPDF==1.26.0
reportlab==4.4.0
presidio-analyzer==2.2.357
presidio-anonymizer==2.2.357
spacy==3.8.5
pytest==8.3.5
